extract Membership source class
Split Source::Membership out into its own file
